+++ __test__/src/tokio.rs
@@ -1,32 +1,33 @@
-pub mod async_std;
-
-pub mod tokio;
-
 #[cfg(test)]
 mod tests {
     use std::{env, fs, path::PathBuf};
 
     use filego::{
-        check::{Check, CheckResult, CheckResultErrorType},
-        merge::Merge,
-        split::{Split, SplitResult},
+        check::{
+            tokio::AsyncCheckExt as _, Check, CheckResult, CheckResultErrorType,
+        },
+        merge::{tokio::AsyncMergeExt as _, Merge},
+        split::{tokio::AsyncSplitExt as _, Split, SplitResult},
     };
 
-    fn setup(cache_name: &str) -> (PathBuf, PathBuf, PathBuf, SplitResult) {
+    async fn setup(
+        cache_name: &str
+    ) -> (PathBuf, PathBuf, PathBuf, SplitResult) {
         let root: PathBuf = env::current_dir().unwrap();
         let file_name: &str = "test.png";
         let chunk_size: usize = 1024 * 1024;
 
         let asset_path: PathBuf = root.join("assets").join(file_name);
         let cache_dir: PathBuf =
-            root.join(".media").join("cache").join("std").join(cache_name);
+            root.join(".media").join("cache").join("tokio").join(cache_name);
 
         // split file
         let split_result: SplitResult = Split::new()
             .in_file(&asset_path)
             .out_dir(&cache_dir)
             .chunk_size(chunk_size)
-            .run()
+            .run_async()
+            .await
             .unwrap();
 
         (
@@ -34,7 +35,7 @@ mod tests {
             cache_dir,
             root.join(".media")
                 .join("output")
-                .join("std")
+                .join("tokio")
                 .join(cache_name)
                 .join(file_name),
             split_result,
@@ -43,7 +44,7 @@ mod tests {
 
     #[tokio::test]
     async fn test_split_file_creates_chunks() {
-        let (_, cache_dir, _, _) = setup("split_file_creates_chunks");
+        let (_, cache_dir, _, _) = setup("split_file_creates_chunks").await;
 
         let chunk_count: usize =
             fs::read_dir(&cache_dir).unwrap().filter_map(Result::ok).count();
@@ -54,13 +55,14 @@ mod tests {
     #[tokio::test]
     async fn test_check_with_missing_chunks() {
         let (_, cache_dir, _, split_result) =
-            setup("check_with_missing_chunks");
+            setup("check_with_missing_chunks").await;
 
         let check_result: CheckResult = Check::new()
             .in_dir(&cache_dir)
             .file_size(split_result.file_size)
             .total_chunks(split_result.total_chunks + 1)
-            .run()
+            .run_async()
+            .await
             .unwrap();
 
         assert!(
@@ -75,13 +77,15 @@ mod tests {
 
     #[tokio::test]
     async fn test_check_with_size_error() {
-        let (_, cache_dir, _, split_result) = setup("check_with_size_error");
+        let (_, cache_dir, _, split_result) =
+            setup("check_with_size_error").await;
 
         let check_result: CheckResult = Check::new()
             .in_dir(&cache_dir)
             .file_size(split_result.file_size + 1)
             .total_chunks(split_result.total_chunks)
-            .run()
+            .run_async()
+            .await
             .unwrap();
 
         assert!(
@@ -96,13 +100,14 @@ mod tests {
 
     #[tokio::test]
     async fn test_successful_check() {
-        let (_, cache_dir, _, split_result) = setup("successful_check");
+        let (_, cache_dir, _, split_result) = setup("successful_check").await;
 
         let check_result: CheckResult = Check::new()
             .in_dir(&cache_dir)
             .file_size(split_result.file_size)
             .total_chunks(split_result.total_chunks)
-            .run()
+            .run_async()
+            .await
             .unwrap();
 
         assert!(
@@ -113,9 +118,15 @@ mod tests {
 
     #[tokio::test]
     async fn test_merge_creates_output_file() {
-        let (_, cache_dir, output_path, _) = setup("merge_creates_output_file");
+        let (_, cache_dir, output_path, _) =
+            setup("merge_creates_output_file").await;
 
-        Merge::new().in_dir(&cache_dir).out_file(&output_path).run().unwrap();
+        Merge::new()
+            .in_dir(&cache_dir)
+            .out_file(&output_path)
+            .run_async()
+            .await
+            .unwrap();
 
         assert!(
             output_path.exists(),
@@ -127,14 +138,14 @@ mod tests {
     async fn test_merge_on_empty_cache_dir() {
         let root: PathBuf = env::current_dir().unwrap();
         let empty_cache_dir: PathBuf =
-            root.join(".media").join("cache").join("std").join("empty_test");
+            root.join(".media").join("cache").join("tokio").join("empty_test");
 
         fs::create_dir_all(&empty_cache_dir).unwrap();
 
         let output_path: PathBuf = root
             .join(".media")
             .join("output")
-            .join("std")
+            .join("tokio")
             .join("empty_test")
             .join("output.txt");
 
@@ -142,7 +153,8 @@ mod tests {
             Merge::new()
                 .in_dir(&empty_cache_dir)
                 .out_file(&output_path)
-                .run()
+                .run_async()
+                .await
                 .is_err(),
             "Merge should fail with an empty cache directory."
         );