+++ package/src/config/mod.rs
@@ -1,5 +1,0 @@
-/// The default chunk size in bytes.
-pub const CHUNK_SIZE_DEFAULT: usize = 2 * 1024 * 1024;
-
-/// The default maximum size of the buffer capacity in bytes.
-pub const BUFFER_CAPACITY_MAX_DEFAULT: usize = 10 * 1024 * 1024;